Switch own WAV reader; start regression test.

+++ packages/trs80-cassette-reader-js/src/DisplaySamples.ts
@@ -6,9 +6,9 @@ export class DisplaySamples {
     /**
      * Index 0 are the original samples, index 1 are halved, etc.
      */
-    public samplesList: Float32Array[];
+    public samplesList: Int16Array[];
 
-    constructor(samples: Float32Array) {
+    constructor(samples: Int16Array) {
         this.samplesList = [samples];
         this.filterDown();
     }
@@ -20,7 +20,7 @@ export class DisplaySamples {
         while (this.samplesList[this.samplesList.length - 1].length > 500) {
             const samples = this.samplesList[this.samplesList.length - 1];
             const half = Math.floor(samples.length / 2);
-            const down = new Float32Array(half);
+            const down = new Int16Array(half);
             for (let i = 0, j = 0; i < half; i++, j += 2) {
                 down[i] = Math.max(samples[j], samples[j + 1]);
             }
